// src/enums/homeEnum.ts
export enum ToastMessageHome {
  CREATE_TRAINING_SUCCESS_MESSAGE = "Treino criado com sucesso!",
  CREATE_WORKOUT_SUCCESS_MESSAGE = "Exercício criado com sucesso!",
  INSERT_ALL_INPUTS = "É preciso inserir todos os campos!",
}

export enum TitleInput {
  TITLE_WORKOUT = "Nome do treino",
  TITLE_EXERCISE = "Nome do exercício",
  TITLE_WEIGHT = "Peso",
  TITLE_UNIT = "Unidade",
  TITLE_FREQUENCE = "Repetições",
}

export enum ButtonTitle {
  CREATE_TRAINING = "Criar Treino",
  CREATE_EXERCISE = "Criar Exercício",
  CREATED = "Criado!",
}